+++ src/contour/Config.h
@@ -179,6 +179,8 @@ struct TerminalProfile
         InputModeConfig visual;
     } inputModes;
 
+    terminal::StatusDisplayType initialStatusDisplayType = terminal::StatusDisplayType::None;
+
     terminal::Opacity backgroundOpacity; // value between 0 (fully transparent) and 0xFF (fully visible).
     bool backgroundBlur;                 // On Windows 10, this will enable Acrylic Backdrop.